requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
vercel_blob==0.4.2
gunicorn==23.0.0
pymongo==4.15.2